requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
environ==1.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
